+++ app/controllers/services/offers_controller.rb
@@ -3,11 +3,16 @@
 class Services::OffersController < Services::ApplicationController
   def index
     @offers = @service.offers
+
+    if @service.offers_count == 1
+      select_offer(@offers.first)
+      redirect_to service_configuration_path(@service)
+    end
   end
 
   def update
     if offer
-      session[session_key] = { "offer_id" => offer.id }
+      select_offer(offer)
       redirect_to service_configuration_path(@service)
     else
       @offers = @service.offers
@@ -17,6 +22,10 @@ class Services::OffersController < Services::ApplicationController
   end
 
   private
+
+    def select_offer(offer)
+      session[session_key] = { "offer_id" => offer.id }
+    end
 
     def offer
       @offer ||= @service.offers.find_by(iid: offer_params[:offer_id])