+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2018_10_22_113346) do
+ActiveRecord::Schema.define(version: 2018_10_25_095520) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -127,6 +127,7 @@ ActiveRecord::Schema.define(version: 2018_10_22_113346) do
     t.boolean "open_access", default: false
     t.bigint "provider_id"
     t.integer "service_opinion_count", default: 0
+    t.text "contact_emails", default: [], array: true
     t.text "places", null: false
     t.text "languages", null: false
     t.text "dedicated_for", null: false
@@ -139,7 +140,7 @@ ActiveRecord::Schema.define(version: 2018_10_22_113346) do
     t.text "tutorial_url", null: false
     t.text "restrictions", null: false
     t.text "phase", null: false
-    t.text "contact_emails", default: [], array: true
+    t.integer "offers_count", default: 0
     t.index ["description"], name: "index_services_on_description"
     t.index ["owner_id"], name: "index_services_on_owner_id"
     t.index ["provider_id"], name: "index_services_on_provider_id"